src/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "copilot-orchestrator"
version = "0.1.0"
description = "Webhook-driven automation service for GitHub Copilot workflow management"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "Grammy Jiang"}]
keywords = ["github", "copilot", "webhooks", "automation", "orchestration"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlmodel>=0.0.22",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "pyjwt[crypto]>=2.9.0",
  "cryptography>=43.0.0",
  "typer>=0.15.0",
  "rich>=13.9.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "pytest-mock>=3.14.0",
  "respx>=0.22.0",
  "httpx>=0.28.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
  "pre-commit>=4.0.0"
]
postgres = ["asyncpg>=0.30.0", "psycopg[binary]>=3.2.0"]

[project.scripts]
copilot-orchestrator = "app.cli:cli_app"

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:"
]
fail_under = 75
show_missing = true

[tool.coverage.run]
source = ["app"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["sqlmodel.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["app.db.models.*"]
# SQLModel uses table=True which mypy doesn't understand
disable_error_code = ["call-arg"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--tb=short", "-ra"]
markers = [
  "unit: Unit tests (fast, no I/O)",
  "integration: Integration tests (may use test DB)",
  "e2e: End-to-end tests (slow, full stack)",
  "security: Security-related tests"
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["app", "tests"]

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # Pyflakes
  "I", # isort
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
  "S", # flake8-bandit (security)
  "T20" # flake8-print
]
ignore = [
  "E501", # Line too long (handled by formatter)
  "B008", # Function call in default argument
  "S101", # assert statements (allowed in tests)
  "S104" # Binding to all interfaces (intentional for development server)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "S105", # Hardcoded password (test fixtures)
  "S106" # Hardcoded password in function argument (test data)
]
"app/services/auth.py" = [
  "S105" # GitHub OAuth URLs contain "token" (false positive)
]
